+++ src/chain/point_list.cpp
@@ -28,12 +28,12 @@ std::vector<libbitcoin::chain::point>& point_list_cpp(point_list_t point_list) {
     return *static_cast<std::vector<libbitcoin::chain::point>*>(point_list);
 }
 
-point_t point_list_nth(point_list_t point_list, size_t n){
+point_t point_list_nth(point_list_t point_list, uint64_t /*size_t*/ n){
     auto& point_n = point_list_cpp(point_list)[n];
     return &point_n;
 }
 
-size_t point_list_count(point_list_t point_list){
+uint64_t /*size_t*/ point_list_count(point_list_t point_list){
     return point_list_const_cpp(point_list).size();
 }
